refactoring - move code to internal/ path
This project's source code is not mean to be consumed
as a library so guard against unexpected external uses
by moving code into an internal/ path.

Reference: https://pkg.go.dev/cmd/go#hdr-Internal_Directories
