+++ internal/pkg/output/util.go
